Took out some unnecessary files and such
