--- FilmBuffs/ContentViewModel.swift
@@ -1,13 +1,0 @@
-//
-//  ContentViewModel.swift
-//  FilmBuffs
-//
-//  Created by Clay Beal on 11/5/24.
-//  View Model
-//
-//import SwiftUI
-//
-//class ContentViewModel: ObservableObject {
-//    // TODO: i dont really know yet
-//}
-
